checking in
from alison:
Jockie to the Fair - please rewrite the attached version with just the A
and B parts. There are a few note changes as well:Part B Meas. 1 - the
first eighth note should be a D. Same change in measure 3. Last measure
of Part B - the first eighth note should be B, the next group of three
eighth notes should be D-C-A. The reason I name it "various" is that
different traditions have different numbers of measures in the B part -
everything from 8 to 14. You just have to know.